requirements.txt
python-telegram-bot==13.11
selenium==4.10.0
webdriver-manager==3.8.6
python-dotenv==1.0.0
